Implemented datasource arch with hardcoded supported rn
