+++ src/types/dataSource.d.ts
@@ -1,0 +1,3 @@
+export enum dataSourceTypes {
+  FILE = 'file'
+}
